requirements.txt
# Core deep learning
torch>=2.0.0
torchtext>=0.15.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# BLEU evaluation
sacrebleu>=2.3.0

# Optional: indic NLP for better tokenisation of Indian languages
# pip install indic-nlp-library

# Development utilities
scikit-learn>=1.3.0
tqdm>=4.66.0
